separate logger for semantic tokens
